+++ src/tools/playlist-management.ts
@@ -34,7 +34,7 @@ import type {
   RemoveTracksFromPlaylistResponse,
   ReorderPlaylistTrackRequest,
   ReorderPlaylistTrackResponse,
-} from '../types/dto.js';
+} from '../types/index.js';
 import { DEFAULT_VALUES } from '../constants/defaults.js';
 
 // Common pagination schema
